Refactor CollectionNavigator to a base and a collection implementation
